Move wait_for_block_import method.
- move ATTEMPTS_NUMBER and PEERS_THRESHOLD constants and increase their values
