+++ crates/subspace-service/src/sync_from_dsn.rs
@@ -14,7 +14,6 @@ use sc_consensus::import_queue::ImportQueueService;
 use sc_consensus_subspace::archiver::SegmentHeadersStore;
 use sc_network::service::traits::NetworkService;
 use sc_network::NetworkBlock;
-pub use snap_sync::wait_for_block_import;
 use sp_api::ProvideRuntimeApi;
 use sp_blockchain::HeaderBackend;
 use sp_consensus_subspace::SubspaceApi;
